-Created CenterSelectionGUI -Implemented checkUserExists(String userid) method in RemoteDatabaseService and ClientManager to check for userid existence in the database -Implemented getCentersList() method in RemoteDatabaseService and ClientManager to get the list of already registered Monitoring Centers -Implemented the completeRegistrationUser(User user, int centerid) method to complete the registration process of an user -Implemented some methods to check fields validity for the registration process in the ClientManager -Implemented the possibility to go bak to the user registration form without losing text already inserted -Added the necessary PredefinedQuery for the registration

+++ src/main/java/org/climatemonitoring/shared/models/MonitoringCenter.java
@@ -29,7 +29,7 @@ public class MonitoringCenter implements Serializable {
         this.province = province;
     }
 
-    public long getCenterid() {
+    public int getCenterid() {
         return centerid;
     }
